sorting by name added in fsd

+++ JavaFiles/fsd/src/main/java/com/example/angular_java_practice/fsd/services/CustomerService.java
@@ -4,9 +4,11 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import com.example.angular_java_practice.fsd.entity.Customer;
+import com.example.angular_java_practice.fsd.exception.CustomerNotFoundException;
 import com.example.angular_java_practice.fsd.exception.InvalidAgeException;
 import com.example.angular_java_practice.fsd.repository.CustomerRepository;
 
@@ -25,11 +27,29 @@ public class CustomerService {
     public Optional<Customer> get_customer_by_id(Integer id) {
         return csr.findById(id);
     }
-    public void deleteCustomerById(Integer id) {
-        csr.deleteById(id);
+    public Optional<Customer> deleteCustomerById(Integer id) {
+        Optional<Customer> customer = get_customer_by_id(id);
+        Optional <Customer> cus=customer;
+        if (customer.isPresent()) {
+            csr.deleteById(id);
+            return cus;
+        }
+        return customer;
     }
     
-    
+    public Optional<Customer> update_customer(Integer id,String location){
+        Optional<Customer> existingCustomer = get_customer_by_id(id);
+        if (existingCustomer.isPresent()) {
+            Customer customerToUpdate = existingCustomer.get();
+            customerToUpdate.setLocation(location); // update location
+            create_customer(customerToUpdate);
+        } 
+        else {
+            
+            throw new CustomerNotFoundException("Customer with ID " + id + " not found.");
+        }
+        return existingCustomer;
+    }
 
    
    
@@ -38,5 +58,10 @@ public class CustomerService {
    
     public List<Customer> get_all_cus() {
         return csr.findAll();
+    }
+
+    public List<Customer> get_all_cus_sorted_by_name(){
+        List<Customer> c=csr.findAll(Sort.by(Sort.Direction.ASC, "name")); // Use Spring Data JPA's Sort utility
+        return c;
     }
 }
